% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{create_corpus}
\alias{create_corpus}
\title{Title}
\usage{
create_corpus(file_loc, read_large = FALSE)
}
\arguments{
\item{file_loc}{The location of the files}

\item{read_large}{A Binary argument TRUE/FALSE whether the large documents are read}
}
\value{
A dataframe
}
\description{
Title
}
