% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{get_file_details}
\alias{get_file_details}
\title{Get paths and info about PAN-AV docs}
\usage{
get_file_details(file_loc)
}
\arguments{
\item{file_loc}{Location of the PAN-AV files.}
}
\value{
A dataframe
}
\description{
Get paths and info about PAN-AV docs
}
