% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{get_file_locs}
\alias{get_file_locs}
\title{Return file paths of all files in location.}
\usage{
get_file_locs(file_loc)
}
\arguments{
\item{file_loc}{PAN-AV data location}
}
\value{
A vector
}
\description{
Return file paths of all files in location.
}
