% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{Read the files required for the corpus}
\usage{
read_files(file_loc, read_large = FALSE)
}
\arguments{
\item{file_loc}{The filepath of the file to read}
}
\value{
A dataframe.
}
\description{
Read the files required for the corpus
}
