% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_txt}
\alias{read_txt}
\title{Read a txt file}
\usage{
read_txt(file_loc)
}
\arguments{
\item{file_loc}{The filepath of the txt file to read.}
}
\value{
A dataframe.
}
\description{
Read a txt file
}
