% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{create_corpus}
\alias{create_corpus}
\title{Create a corpus object from a folder containing PAN competition data.}
\usage{
create_corpus(
  file_loc,
  selected_comp = NULL,
  read_large = FALSE,
  convert_to_corpus = FALSE
)
}
\arguments{
\item{file_loc}{The location of the files}

\item{selected_comp}{A vector containing competitions to select e.g. c('pan21')}

\item{read_large}{A Binary argument TRUE/FALSE whether the large documents are read}

\item{convert_to_corpus}{A Binary argument TRUE/FALSE to detemine whether to convert to Quanteda corpus object.}
}
\value{
A dataframe or corpus object depending on the user's selection.
}
\description{
Create a corpus object from a folder containing PAN competition data.
}
