% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jsonl.R
\name{read_jsonl}
\alias{read_jsonl}
\title{Read a jsonl file}
\usage{
read_jsonl(file_loc)
}
\arguments{
\item{file_loc}{The location of the jsonl file to read.}
}
\value{
A dataframe.
}
\description{
Read a jsonl file
}
\examples{
# file_loc <- ".data/pan20-authorship-verification-training-example.jsonl"
# usedf <- read_jsonl(file_loc)
}
